# Use `new_tibble()` in `$pick()` for faster construction