``` r
library(dplyr)

r_iris <- rowwise(iris)

bench::mark(
  mutate(r_iris, x = across(Sepal.Width)),
  iterations = 50
)

# master
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression                                min median `itr/sec` mem_alloc
#>   <bch:expr>                              <bch> <bch:>     <dbl> <bch:byt>
#> 1 mutate(r_iris, x = across(Sepal.Width)) 276ms  318ms      3.09     3.4MB
#> # … with 1 more variable: `gc/sec` <dbl>

# this PR
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression                                min median `itr/sec` mem_alloc
#>   <bch:expr>                              <bch> <bch:>     <dbl> <bch:byt>
#> 1 mutate(r_iris, x = across(Sepal.Width)) 218ms  258ms      3.72    2.87MB
#> # … with 1 more variable: `gc/sec` <dbl>
```

<sup>Created on 2020-02-27 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0)</sup>